+++ vesuvius/src/vesuvius/neural_tracing/fiber_trace_3d/__init__.py
@@ -1,6 +1,7 @@
 """CP-centered 3D fiber tracing training path."""
 
 from .direction import (
+    decode_lasagna_direction_3x2_analytic,
     decode_lasagna_direction_2d,
     encode_lasagna_direction_2d,
     encode_lasagna_direction_3x2,
@@ -10,6 +11,7 @@ from .model import FiberTrace3DModelConfig, FiberTrace3DNet
 __all__ = [
     "FiberTrace3DModelConfig",
     "FiberTrace3DNet",
+    "decode_lasagna_direction_3x2_analytic",
     "decode_lasagna_direction_2d",
     "encode_lasagna_direction_2d",
     "encode_lasagna_direction_3x2",